feat(logger): Allow logger to be set in config
Previously configuration would happen _before_ the client.logger() method could be called, so the
"[bugsnag] loaded!" message would be printed before the logging could be disabled. This commit
allows the logger to be set in the configuration object. Logging can be completely disabled by
passing `logger: null`.

Fixes #302
